Move to diagnostics module
